+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # Changelog
 
-## 4.3.4 / Unreleased
+## 4.4.0 / Unreleased
 
 - Test with Python 3.13
 - Use ruff instead of black/isort
 - [#343](https://github.com/mar10/wsgidav/issues/343) 
   Add experimental middleware for impersonation (@leo9800)
-- [New option `default_charset`](https://github.com/mar10/wsgidav/pull/354)
-  Defaults to 'utf-8' (@leo9800)
+- [#352](https://github.com/mar10/wsgidav/pull/352)
+  Apache htpasswd based domain controller (@leo9800)
+- [#354`](https://github.com/mar10/wsgidav/pull/354)
+  New option `default_charset, defaults to 'utf-8' (@leo9800)
 
 ## 4.3.3 / 2024-05-04